Moved StreamError to IceRpc.Internal

+++ src/IceRpc/Slice/AsyncEnumerableStreamParamSender.cs
@@ -109,7 +109,7 @@ namespace IceRpc.Slice
             }
             catch
             {
-                multiplexedStream.AbortWrite(StreamError.StreamingCanceledByWriter);
+                multiplexedStream.AbortWrite((byte)StreamError.StreamingCanceledByWriter);
                 throw;
             }
             finally
